refactor(coalescent): move log-to-probability bridge into distribution layer
The coalescent model owned the grid walk, exp(min - cost) scaling, and normalization that turned per-time costs into a probability distribution. Extract that as distribution_apply_neg_log_weight in treetime-distribution; the model now only supplies per-time contributions. Behavior is unchanged: the same cost-min-shifted exponential and peak normalization run on the same grid.
